merge main to jeff's dev branch

+++ dacapo/predict.py
@@ -24,7 +24,7 @@ def predict(
     num_cpu_workers: int = 4,
     compute_context: ComputeContext = LocalTorch(),
     output_roi: Optional[Roi] = None,
-    output_dtype: Optional[np.dtype] = np.uint8,
+    output_dtype: Optional[np.dtype] = np.float32,  # add necessary type conversions
     overwrite: bool = False,
 ):
     # get the model's input and output size
